+++ packages/eslint-plugin-template/src/rules/prefer-at-empty.ts
@@ -621,10 +621,7 @@ interface ForNodeInfo {
 interface IfNodeInfo {
   readonly node: TmplAstIfBlock;
   readonly kind:
-    | 'if-empty'
-    | 'if-empty-else'
-    | 'if-not-empty'
-    | 'if-not-empty-else';
+    'if-empty' | 'if-empty-else' | 'if-not-empty' | 'if-not-empty-else';
   readonly collection: AST;
 }