+++ packages/eslint-plugin-template/src/rules/prefer-contextual-for-variables.ts
@@ -719,8 +719,7 @@ interface ForLoopInfo {
   readonly variables: readonly TmplAstVariable[];
   readonly references: Map<string, TSESTree.Range[]> | undefined;
   simplifications?:
-    | Partial<Record<SimplificationMessageIds, Simplification[]>>
-    | undefined;
+    Partial<Record<SimplificationMessageIds, Simplification[]>> | undefined;
 }
 
 interface Simplifications {